+++ scripts/build.sh
@@ -2,8 +2,17 @@
 
 set -e
 
-cd `dirname $0`/..
+HERE=`dirname $0`
+cd $HERE/..
 ROOT=`pwd`
+
+CFGFILE=${ROOT}/scripts/build.cfg
+
+if [ -e $CFGFILE ]; then
+    source $CFGFILE
+fi
+
+FLATCC_BUILD_CMD=${FLATCC_BUILD_CMD:-ninja}
 
 mkdir -p ${ROOT}/bin
 mkdir -p ${ROOT}/lib
@@ -17,7 +26,7 @@ if [ ! -d ${ROOT}/build/Debug ] || [ ! -d  ${ROOT}/build/Release ]; then
 fi
 
 echo "building Debug" 1>&2
-cd ${ROOT}/build/Debug && ninja
+cd ${ROOT}/build/Debug && $FLATCC_BUILD_CMD
 
 echo "building Release" 1>&2
-cd ${ROOT}/build/Release && ninja
+cd ${ROOT}/build/Release && $FLATCC_BUILD_CMD